refactor(sections-editor): extract array-item row renderers into their own module
ArrayField's ArrayRowContent and SortableArrayRow are pure, props-only
presentational components used for the drag overlay and sortable list rows.
Move them to fields/array-row.tsx to shrink array-field.tsx and make the row
UI independently testable/reusable, with no behavior change.
